+++ CHANGELOG.md
@@ -13,11 +13,13 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 # Added
 
 - The `devices` command is added. Included are:
-    - `devices deactivate` to deactivate a single computer.
-    - `devices show` to retrieve detailed information about a computer.
+    - `devices deactivate` to deactivate a single device.
+    - `devices reactivate` to reactivate a single device.
+    - `devices show` to retrieve detailed information about a device.
     - `devices list` to retrieve info about many devices, including device settings.
     - `devices list-backup-sets` to retrieve detailed info about device backup sets.
     - `devices bulk deactivate` to deactivate a list of devices.
+    - `devices bulk reactivate` to reactivate a list of devices.
 
 - `code42 departing-employee list` command.